Fix signed / unsigned comparison warning [Fixes #60]
This iterator value is used when looping through a hash with #each to
keep track of where the iterator is. Size() of a Hash can only be a
positive value.

+++ rice/Hash.hpp
@@ -199,7 +199,7 @@ protected:
 
 private:
   Hash hash_;
-  int current_index_;
+  size_t current_index_;
   VALUE keys_;
 
   mutable typename detail::remove_const<Value_T>::Type tmp_;
